Move the duplicate api to workspace plugin
