+++ src/core/server/saved_objects/routes/index.ts
@@ -45,7 +45,6 @@ import { registerExportRoute } from './export';
 import { registerImportRoute } from './import';
 import { registerResolveImportErrorsRoute } from './resolve_import_errors';
 import { registerMigrateRoute } from './migrate';
-import { registerCopyRoute } from './copy';
 
 export function registerRoutes({
   http,
@@ -72,7 +71,6 @@ export function registerRoutes({
   registerExportRoute(router, config);
   registerImportRoute(router, config);
   registerResolveImportErrorsRoute(router, config);
-  registerCopyRoute(router, config);
 
   const internalRouter = http.createRouter('/internal/saved_objects/');
